requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncmy==0.2.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
exceptiongroup==1.2.2
Faker==36.1.1
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
mysql-connector-python==9.2.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
